Use complete base name for quick ID's
This is risky for probably having wrong assets for ROM's with a bad
naming scheme eg. same ROM name but different region.

Benefit is that moving ROM's is easier as there's no path involved
anymore.
The same applies for converting ROM archives to a different format eg.
TorrentZip to Torrent7z or to CHD's since the file extension is now
ignored also.